+++ bioprov/src/files.py
@@ -53,7 +53,7 @@ class File:
         self.attributes = attributes
         self._exists = self.path.exists()
         self.size = get_size(self.path)
-        self.raw_size = get_size(self.path, convert=False)
+        self._raw_size = get_size(self.path, convert=False)
         self._sha256 = file_to_sha256(self.path)
 
         # Provenance attributes
@@ -97,6 +97,15 @@ class File:
     @size.setter
     def size(self, value):
         self._size = value
+
+    @property
+    def raw_size(self):
+        self._raw_size = get_size(self.path, convert=False)
+        return self._raw_size
+
+    @raw_size.setter
+    def raw_size(self, value):
+        self._raw_size = value
 
     @property
     def entity(self):