+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Fix bug on JSON import when documents are in the Project [ ]
 * Fix 'Too many open files' bug, when reading TempFile when loading Project [ ]
 * Fix `File.size` property [x]
+* Add `File.raw_size` property [x]
 
 ### v0.1.23
 * Patch PresetProgram SeqFile addition feature [x]